Split nested_growables into one file per type
